change int64 instead of int for transplanting

+++ webapp/go/livecomment_handler.go
@@ -16,44 +16,44 @@ import (
 
 type PostLivecommentRequest struct {
 	Comment string `json:"comment"`
-	Tip     int    `json:"tip"`
+	Tip     int64  `json:"tip"`
 }
 
 type LivecommentModel struct {
-	Id           int       `db:"id"`
-	UserId       int       `db:"user_id"`
-	LivestreamId int       `db:"livestream_id"`
+	Id           int64     `db:"id"`
+	UserId       int64     `db:"user_id"`
+	LivestreamId int64     `db:"livestream_id"`
 	Comment      string    `db:"comment"`
-	Tip          int       `db:"tip"`
-	ReportCount  int       `db:"report_count"`
+	Tip          int64     `db:"tip"`
+	ReportCount  int64     `db:"report_count"`
 	CreatedAt    time.Time `db:"created_at"`
 	UpdatedAt    time.Time `db:"updated_at"`
 }
 
 type Livecomment struct {
-	Id          int        `json:"id"`
+	Id          int64      `json:"id"`
 	User        User       `json:"user"`
 	Livestream  Livestream `json:"livestream"`
 	Comment     string     `json:"comment"`
-	Tip         int        `json:"tip"`
-	ReportCount int        `json:"report_count"`
-	CreatedAt   int        `json:"created_at"`
-	UpdatedAt   int        `json:"updated_at"`
+	Tip         int64      `json:"tip"`
+	ReportCount int64      `json:"report_count"`
+	CreatedAt   int64      `json:"created_at"`
+	UpdatedAt   int64      `json:"updated_at"`
 }
 
 type LivecommentReport struct {
-	Id          int         `json:"id"`
+	Id          int64       `json:"id"`
 	Reporter    User        `json:"reporter"`
 	Livecomment Livecomment `json:"livecomment"`
-	CreatedAt   int         `json:"created_at"`
-	UpdatedAt   int         `json:"updated_at"`
+	CreatedAt   int64       `json:"created_at"`
+	UpdatedAt   int64       `json:"updated_at"`
 }
 
 type LivecommentReportModel struct {
-	Id            int       `db:"id"`
-	UserId        int       `db:"user_id"`
-	LivestreamId  int       `db:"livestream_id"`
-	LivecommentId int       `db:"livecomment_id"`
+	Id            int64     `db:"id"`
+	UserId        int64     `db:"user_id"`
+	LivestreamId  int64     `db:"livestream_id"`
+	LivecommentId int64     `db:"livecomment_id"`
 	CreatedAt     time.Time `db:"created_at"`
 	UpdatedAt     time.Time `db:"updated_at"`
 }
@@ -63,8 +63,8 @@ type ModerateRequest struct {
 }
 
 type NGWord struct {
-	UserId       int    `json:"user_id" db:"user_id"`
-	LivestreamId int    `json:"livestream_id" db:"livestream_id"`
+	UserId       int64  `json:"user_id" db:"user_id"`
+	LivestreamId int64  `json:"livestream_id" db:"livestream_id"`
 	Word         string `json:"word" db:"word"`
 }
 
@@ -160,8 +160,8 @@ func postLivecommentHandler(c echo.Context) error {
 
 	now := time.Now()
 	livecommentModel := LivecommentModel{
-		UserId:       userId,
-		LivestreamId: livestreamId,
+		UserId:       int64(userId),
+		LivestreamId: int64(livestreamId),
 		Comment:      req.Comment,
 		Tip:          req.Tip,
 		CreatedAt:    now,
@@ -177,7 +177,7 @@ func postLivecommentHandler(c echo.Context) error {
 	if err != nil {
 		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
 	}
-	livecommentModel.Id = int(livecommentId)
+	livecommentModel.Id = livecommentId
 
 	livecomment, err := fillLivecommentResponse(ctx, tx, livecommentModel)
 	if err != nil {
@@ -234,9 +234,9 @@ func reportLivecommentHandler(c echo.Context) error {
 
 	now := time.Now()
 	reportModel := LivecommentReportModel{
-		UserId:        userId,
-		LivestreamId:  livestreamId,
-		LivecommentId: livecommentId,
+		UserId:        int64(userId),
+		LivestreamId:  int64(livestreamId),
+		LivecommentId: int64(livecommentId),
 		CreatedAt:     now,
 		UpdatedAt:     now,
 	}
@@ -253,7 +253,7 @@ func reportLivecommentHandler(c echo.Context) error {
 	if err != nil {
 		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
 	}
-	reportModel.Id = int(reportId)
+	reportModel.Id = reportId
 
 	report, err := fillLivecommentReportResponse(ctx, tx, reportModel)
 	if err != nil {
@@ -309,8 +309,8 @@ func moderateNGWordHandler(c echo.Context) error {
 	}
 
 	rs, err := tx.NamedExecContext(ctx, "INSERT INTO ng_words(user_id, livestream_id, word) VALUES (:user_id, :livestream_id, :word)", &NGWord{
-		UserId:       userId,
-		LivestreamId: livestreamId,
+		UserId:       int64(userId),
+		LivestreamId: int64(livestreamId),
 		Word:         req.NGWord,
 	})
 	if err != nil {
@@ -357,8 +357,8 @@ func fillLivecommentResponse(ctx context.Context, tx *sqlx.Tx, livecommentModel 
 		Comment:     livecommentModel.Comment,
 		Tip:         livecommentModel.Tip,
 		ReportCount: livecommentModel.ReportCount,
-		CreatedAt:   int(livecommentModel.CreatedAt.Unix()),
-		UpdatedAt:   int(livecommentModel.UpdatedAt.Unix()),
+		CreatedAt:   livecommentModel.CreatedAt.Unix(),
+		UpdatedAt:   livecommentModel.UpdatedAt.Unix(),
 	}
 
 	return livecomment, nil
@@ -387,8 +387,8 @@ func fillLivecommentReportResponse(ctx context.Context, tx *sqlx.Tx, reportModel
 		Id:          reportModel.Id,
 		Reporter:    reporter,
 		Livecomment: livecomment,
-		CreatedAt:   int(reportModel.CreatedAt.Unix()),
-		UpdatedAt:   int(reportModel.UpdatedAt.Unix()),
+		CreatedAt:   reportModel.CreatedAt.Unix(),
+		UpdatedAt:   reportModel.UpdatedAt.Unix(),
 	}
 	return report, nil
 }
