+++ webapp/go/reaction_handler.go
@@ -13,19 +13,19 @@ import (
 )
 
 type ReactionModel struct {
-	Id           int       `db:"id"`
+	Id           int64     `db:"id"`
 	EmojiName    string    `db:"emoji_name"`
-	UserId       int       `db:"user_id"`
-	LivestreamId int       `db:"livestream_id"`
+	UserId       int64     `db:"user_id"`
+	LivestreamId int64     `db:"livestream_id"`
 	CreatedAt    time.Time `db:"created_at"`
 }
 
 type Reaction struct {
-	Id         int        `json:"id"`
+	Id         int64      `json:"id"`
 	EmojiName  string     `json:"emoji_name"`
 	User       User       `json:"user"`
 	Livestream Livestream `json:"livestream"`
-	CreatedAt  int        `json:"created_at"`
+	CreatedAt  int64      `json:"created_at"`
 }
 
 type PostReactionRequest struct {
@@ -104,8 +104,8 @@ func postReactionHandler(c echo.Context) error {
 	defer tx.Rollback()
 
 	reactionModel := ReactionModel{
-		UserId:       userId,
-		LivestreamId: livestreamId,
+		UserId:       int64(userId),
+		LivestreamId: int64(livestreamId),
 		EmojiName:    req.EmojiName,
 	}
 
@@ -118,7 +118,7 @@ func postReactionHandler(c echo.Context) error {
 	if err != nil {
 		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
 	}
-	reactionModel.Id = int(reactionId)
+	reactionModel.Id = reactionId
 
 	reaction, err := fillReactionResponse(ctx, tx, reactionModel)
 	if err != nil {
@@ -156,7 +156,7 @@ func fillReactionResponse(ctx context.Context, tx *sqlx.Tx, reactionModel Reacti
 		EmojiName:  reactionModel.EmojiName,
 		User:       user,
 		Livestream: livestream,
-		CreatedAt:  int(reactionModel.CreatedAt.Unix()),
+		CreatedAt:  reactionModel.CreatedAt.Unix(),
 	}
 
 	return reaction, nil
